+++ src/app/project-manager/project-manager-home/project-manager-home.component.html
@@ -1,6 +1,6 @@
 <div class="photo-card">
     <img
-      src="\assets\PM3.jpeg"
+      src="\assets\team-member.png"
       alt="Welcome Image"
       class="photo-card-img"
     />
