+++ src/app/project-manager/project-manager-home/project-manager-home.component.scss
@@ -17,18 +17,16 @@
         left: 50%;
         transform: translate(-50%, -50%);
         text-align: center;
-        color: #1C1F2A;
+        color: white;
         padding: 2vh 5%;
         border-radius: 10px;
         h1 {
             font-size: 5vh;
-            margin-bottom: 5%;
+            margin-bottom: 2vh;
         }
 
         p {
             font-size: 2.5vh;
-            margin-bottom: 10%;
-
         }
     }
 }
