+++ src/app/project-manager/project-manager-projects/project-manager-projects.component.ts
@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, FormArray } from '@angular/forms';
 import { ReactiveFormsModule } from '@angular/forms';
-import { NgFor } from '@angular/common';
 
 @Component({
   selector: 'app-project-form',
-  imports:[ReactiveFormsModule, NgFor] ,
+  imports:[ReactiveFormsModule] ,
   templateUrl: './project-manager-projects.component.html',
   styleUrls: ['./project-manager-projects.component.scss']
