--- src/app/project-manager/project-manager-tasks/create-tasks/create-tasks.component.scss
@@ -1,92 +1,0 @@
-:host {
-  display: block;
-  padding: 2rem;
-  max-width: 800px;
-  margin: 0 auto;
-  font-family: 'Segoe UI', sans-serif;
-}
-
-form {
-  background-color: #fff;
-  padding: 2rem;
-  border-radius: 1rem;
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.06);
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-label {
-  display: flex;
-  flex-direction: column;
-  font-weight: 600;
-  font-size: 0.95rem;
-  color: #333;
-  gap: 0.5rem;
-}
-
-input,
-textarea {
-  padding: 0.75rem 1rem;
-  font-size: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 0.5rem;
-  transition: border-color 0.2s ease, box-shadow 0.2s ease;
-
-  &:focus {
-    outline: none;
-    border-color: #007bff;
-    box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.2);
-  }
-}
-
-textarea {
-  resize: vertical;
-  min-height: 120px;
-}
-
-.form-array {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-
-  label {
-    font-size: 1rem;
-    margin-bottom: 0.25rem;
-  }
-
-  input {
-    margin-bottom: 0.5rem;
-  }
-}
-
-button[type="button"].add-button {
-  background-color: #f8f9fa;
-  color: #333;
-  border: 1px solid #ccc;
-  padding: 0.5rem 1rem;
-  border-radius: 0.4rem;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-
-  &:hover {
-    background-color: #e2e6ea;
-  }
-}
-
-button[type="submit"] {
-  background-color: #ffa500;
-  color: #fff;
-  font-weight: 600;
-  font-size: 1rem;
-  padding: 0.75rem 1.5rem;
-  border: none;
-  border-radius: 0.5rem;
-  cursor: pointer;
-  align-self: flex-start;
-  transition: background-color 0.2s ease;
-
-  &:hover {
-    background-color: #956203;
-  }
-}
