--- src/app/project-manager/project-manager-tasks/create-tasks/create-tasks.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CreateTasksComponent } from './create-tasks.component';
-
-describe('CreateTasksComponent', () => {
-  let component: CreateTasksComponent;
-  let fixture: ComponentFixture<CreateTasksComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [CreateTasksComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(CreateTasksComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
