--- src/app/project-manager/project-manager-tasks/edit-tasks/edit-tasks.component.html
@@ -1,1 +1,0 @@
-<p>edit-tasks works!</p>
