--- src/app/project-manager/project-manager-tasks/edit-tasks/edit-tasks.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-edit-tasks',
-  imports: [],
-  templateUrl: './edit-tasks.component.html',
-  styleUrl: './edit-tasks.component.scss'
-})
-export class EditTasksComponent {
-
-}
