+++ src/app/project-manager/project-manager-tasks/project-manager-tasks.component.scss
@@ -1,173 +1,92 @@
-// :host {
-//   display: block;
-//   padding: 2rem;
-//   max-width: 800px;
-//   margin: 0 auto;
-//   font-family: 'Segoe UI', sans-serif;
-// }
-
-// form {
-//   background-color: #fff;
-//   padding: 2rem;
-//   border-radius: 1rem;
-//   box-shadow: 0 8px 24px rgba(0, 0, 0, 0.06);
-//   display: flex;
-//   flex-direction: column;
-//   gap: 1.5rem;
-// }
-
-// label {
-//   display: flex;
-//   flex-direction: column;
-//   font-weight: 600;
-//   font-size: 0.95rem;
-//   color: #333;
-//   gap: 0.5rem;
-// }
-
-// input,
-// textarea {
-//   padding: 0.75rem 1rem;
-//   font-size: 1rem;
-//   border: 1px solid #ccc;
-//   border-radius: 0.5rem;
-//   transition: border-color 0.2s ease, box-shadow 0.2s ease;
-
-//   &:focus {
-//     outline: none;
-//     border-color: #007bff;
-//     box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.2);
-//   }
-// }
-
-// textarea {
-//   resize: vertical;
-//   min-height: 120px;
-// }
-
-// .form-array {
-//   display: flex;
-//   flex-direction: column;
-//   gap: 0.75rem;
-
-//   label {
-//     font-size: 1rem;
-//     margin-bottom: 0.25rem;
-//   }
-
-//   input {
-//     margin-bottom: 0.5rem;
-//   }
-// }
-
-// button[type="button"].add-button {
-//   background-color: #f8f9fa;
-//   color: #333;
-//   border: 1px solid #ccc;
-//   padding: 0.5rem 1rem;
-//   border-radius: 0.4rem;
-//   cursor: pointer;
-//   transition: background-color 0.2s ease;
-
-//   &:hover {
-//     background-color: #e2e6ea;
-//   }
-// }
-
-// button[type="submit"] {
-//   background-color: #ffa500;
-//   color: #fff;
-//   font-weight: 600;
-//   font-size: 1rem;
-//   padding: 0.75rem 1.5rem;
-//   border: none;
-//   border-radius: 0.5rem;
-//   cursor: pointer;
-//   align-self: flex-start;
-//   transition: background-color 0.2s ease;
-
-//   &:hover {
-//     background-color: #956203;
-//   }
-// }
-.tasks-page {
+:host {
+  display: block;
   padding: 2rem;
-  background-color: #f9f9f9;
-  min-height: 100vh;
+  max-width: 800px;
+  margin: 0 auto;
+  font-family: 'Segoe UI', sans-serif;
+}
 
-  .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 1.5rem;
+form {
+  background-color: #fff;
+  padding: 2rem;
+  border-radius: 1rem;
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.06);
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
 
-    h1 {
-      font-size: 2rem;
-      color: #333;
-    }
+label {
+  display: flex;
+  flex-direction: column;
+  font-weight: 600;
+  font-size: 0.95rem;
+  color: #333;
+  gap: 0.5rem;
+}
 
-    .create-btn {
-      background-color: #ffa500;
-      color: white;
-      padding: 0.5rem 1rem;
-      border: none;
-      border-radius: 4px;
-      font-weight: 500;
-      cursor: pointer;
+input,
+textarea {
+  padding: 0.75rem 1rem;
+  font-size: 1rem;
+  border: 1px solid #ccc;
+  border-radius: 0.5rem;
+  transition: border-color 0.2s ease, box-shadow 0.2s ease;
 
-      &:hover {
-        background-color: #956203;
-      }
-    }
+  &:focus {
+    outline: none;
+    border-color: #007bff;
+    box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.2);
+  }
+}
+
+textarea {
+  resize: vertical;
+  min-height: 120px;
+}
+
+.form-array {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+
+  label {
+    font-size: 1rem;
+    margin-bottom: 0.25rem;
   }
 
-  .tasks-table {
-    width: 100%;
-    border-collapse: collapse;
-    background: white;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+  input {
+    margin-bottom: 0.5rem;
+  }
+}
 
-    th, td {
-      padding: 1rem;
-      text-align: left;
-      border-bottom: 1px solid #e0e0e0;
-    }
+button[type="button"].add-button {
+  background-color: #f8f9fa;
+  color: #333;
+  border: 1px solid #ccc;
+  padding: 0.5rem 1rem;
+  border-radius: 0.4rem;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
 
-    th {
-      background-color: #1C1F2A;
-      color: white;
-      font-weight: 600;
-    }
+  &:hover {
+    background-color: #e2e6ea;
+  }
+}
 
-    .actions {
-      display: flex;
-      gap: 0.5rem;
-    }
+button[type="submit"] {
+  background-color: #ffa500;
+  color: #fff;
+  font-weight: 600;
+  font-size: 1rem;
+  padding: 0.75rem 1.5rem;
+  border: none;
+  border-radius: 0.5rem;
+  cursor: pointer;
+  align-self: flex-start;
+  transition: background-color 0.2s ease;
 
-    .btn {
-      padding: 0.4rem 0.8rem;
-      border: none;
-      border-radius: 4px;
-      cursor: pointer;
-      font-size: 0.9rem;
-
-      &.edit-btn {
-        background-color: #ffc107;
-        color: #fff;
-
-        &:hover {
-          background-color: #e0a800;
-        }
-      }
-
-      &.delete-btn {
-        background-color: #dc3545;
-        color: #fff;
-
-        &:hover {
-          background-color: #c82333;
-        }
-      }
-    }
+  &:hover {
+    background-color: #956203;
   }
 }
