+++ src/app/project-manager/project-manager.component.scss
@@ -1,19 +1,14 @@
 $arrow-transition-duration: 400ms;
 @use '@angular/material' as mat;
-$dark-grey-color: #1C1F2A;
-$orange-color: #FE7F0E;
 :host{
+    @include mat.elevation-classes();
+    @include mat.app-background();
     @include mat.sidenav-overrides((
-        container-text-color: white,
-        container-background-color: $dark-grey-color
-    ));
-    @include mat.button-overrides((
-        text-label-text-color: white
+        container-text-color: #ffa500,
     ));
     width: 100%;
     height: 100%;
 }
-
 
 button{
     margin-left: 3px;
@@ -27,10 +22,6 @@ label{
     padding: 2px;
 }
 .active{
-    background-color: $orange-color;
-    @include mat.button-overrides((
-    text-label-text-color: white  
-    ));
 }
 .isOpened{
     transform: rotate(180deg);
@@ -39,27 +30,4 @@ label{
 .isClosed{
     transform: rotate(0deg);
     transition-duration: $arrow-transition-duration;
-}
-.horizontal-link {
-    display: flex;
-    font-size: 1.5em;
-    align-items: center;
-    justify-content:center;
-}
-  
-.icon {
-    display: inline-block;
-    width: 5vh;
-    text-align: center;
-    margin-right: 0.2em;
-}
-
-.logout-section{
-    display: flex;
-    cursor: pointer;
-    align-items: center;
-    justify-content:center;
-    width: 100%;
-    position: absolute;
-    bottom: 0px;
 }
