+++ src/app/project-manager/project-manager.component.ts
@@ -1,5 +1,5 @@
 import { Component ,inject } from '@angular/core';
-import { RouterLink, RouterLinkActive, RouterOutlet, Router } from '@angular/router';
+import { RouterLink, RouterLinkActive, RouterOutlet } from '@angular/router';
 
 import { MatButtonModule } from '@angular/material/button';
 import { MatSidenavModule } from '@angular/material/sidenav';
@@ -15,15 +15,10 @@ import { ProjectManagerService } from './project-manager.service';
 })
 
 export class ProjectManagerComponent {
-  private membersService = inject(ProjectManagerService);
+  private membersService=inject(ProjectManagerService);
   isOpened: boolean = false;
-  private router=inject(Router);
   projects=this.membersService.getProjectsByUserId(this.membersService.loggedInUser()?.userID || -1);
   toggleDrawer(): void {
     this.isOpened = !this.isOpened;
 }
-  logout() {
-    this.membersService.logout(); 
-    this.router.navigate(['/']); 
-  }
 }
