+++ src/app/project-manager/project-manager.service.ts
@@ -1,8 +1,6 @@
 import { Injectable, signal } from '@angular/core';
 import { dummyTeamMembers, dummyProjects, dummyTasks } from '../members/dummy-members';
 import { dummyProjectManager } from '../members/dummy-members';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
 import {
   type User,
   type Project,
@@ -15,7 +13,7 @@ import {
   providedIn: 'root',
 })
 export class ProjectManagerService {
-  constructor(private http: HttpClient) {}
+  private readonly teamMembers = dummyTeamMembers;
   private readonly projectManagers = dummyProjectManager;
   private readonly projects = dummyProjects;
   private readonly tasks = dummyTasks;
@@ -23,26 +21,10 @@ export class ProjectManagerService {
   private loggedInUserWritableSignal = signal<User | null>(null);
   loggedInUser = this.loggedInUserWritableSignal.asReadonly();
 
-// getProjectByProjectId(id: number): Observable<Project> {
-//   return this.http.get<Project>(`http://localhost:8060/ProjectManager/getProject?id=${id}`);
-// }
-//   getAllProjects(): Observable<Project[]> {
-//   return this.http.get<Project[]>(`http://localhost:8060/project/getAll`);
-// }
-//   logIn(user: User){
-//     this.loggedInUserWritableSignal.set(user);
-//   }
-
- isUserLoggedIn(): boolean {
-    if (this.loggedInUser()) {
-      return true;
-    } else {
-      return false;
-    }
+  getProjectByProjectId(id: number) {
+    return this.projects.find((project) => project.projectID === id);
   }
-    logout(): void {
-    this.loggedInUserWritableSignal.set(null);
-    }
+
   getProjectsByUserId(userId: number): Project[] {
     return this.projects.filter((project) =>
       project.createdBy.userID
@@ -63,5 +45,13 @@ export class ProjectManagerService {
       (member) => member.userID === user?.userID
     );
   }
+
+  getSubmissionTasksForLoggedInUser(): Task[] {
+    return this.tasks.filter((task) =>
+      this.isUserAssignedInTask(this.loggedInUser(), task)
+    );
+  }
+
+
 
 }
