+++ src/app/project-manager/project-manger.routes.ts
@@ -7,8 +7,6 @@ import { ProjectManagerService } from "./project-manager.service";
 import { ProjectManagerTasksComponent } from "./project-manager-tasks/project-manager-tasks.component";
 import { ProjectManagerTeamsComponent } from "./project-manager-teams/project-manager-teams.component";
 import { ProjectManagerProjectsComponent } from "./project-manager-projects/project-manager-projects.component";
-import { CreateTasksComponent } from "./project-manager-tasks/create-tasks/create-tasks.component";
-import { EditTasksComponent } from "./project-manager-tasks/edit-tasks/edit-tasks.component";
   
 
 type ResolveFn<T> = (  route: ActivatedRouteSnapshot,  state: RouterStateSnapshot) => MaybeAsync<T | RedirectCommand>
@@ -20,15 +18,15 @@ export const ProjectManagerRoutes: Routes=[
     {
         pathMatch:'full',
         path: '',
-        redirectTo: '/homepage'
+        redirectTo: 'homepage'
     },
     {
         path: 'homepage',
         component: ProjectManagerHomeComponent
     },
      {
-        path:'projects',
-        component:ProjectManagerProjectsComponent
+            path:'projects',
+            component:ProjectManagerProjectsComponent
     },
     {
         path:'tasks',
@@ -37,14 +35,6 @@ export const ProjectManagerRoutes: Routes=[
     {
         path:'teams',
         component:ProjectManagerTeamsComponent
-    },
-    {
-        path:'createTasks',
-        component:CreateTasksComponent
-    },
-    {
-        path:'editTasks',
-        component:EditTasksComponent
     }
